Remove inaccurate test
Unintentionally checked this test in when it was meant to be
experimental.

The reason the test fails is that the Google Analytics ID is stored in
an external script and I'd probably have to mock the GA call to properly
test it; I'm too lazy to figure that out right now.